[FORK-REFACTOR] Externalize AES-GCM logic to crypto-clever.ts
Sprint 2 commit `1746dfc950` (SEC-100) rewrote `packages/lib/crypto.ts`
in place to swap AES-256-CBC for AES-256-GCM. That left a ~40-line
diff against `upstream/main`, which means every upstream change to
crypto.ts will land us in a manual merge.

Move the implementation to `packages/lib/crypto-clever.ts`. The
upstream-facing file is now a 3-line shim that re-exports the v2
functions under the upstream names. New `git diff upstream/main`
on crypto.ts: 3 insertions, 41 deletions (the deletions are
mechanical — any future upstream change to the deleted block will
fail the rebase loudly, which is exactly what we want).

Test file split mirrors the implementation:
  - crypto.test.ts restored to upstream content (modulo a tiny
    relaxation of the format assertion that's now CBC-specific)
  - crypto-clever.test.ts holds the 18 v2-specific tests (GCM
    round-trip, auth-tag tamper detection, ciphertext tamper
    detection, legacy CBC fallback, isLegacyCiphertext helper)

All 28 tests pass.
